cleanup harness project files, add test/install
