--- workers/ruby/Steepfile
@@ -1,20 +1,0 @@
-# frozen_string_literal: true
-
-D = Steep::Diagnostic
-
-common_diagnostics = {
-  D::Ruby::UnknownConstant => :information,
-  D::Ruby::UnannotatedEmptyCollection => :information,
-  D::Ruby::UndeclaredMethodDefinition => :information
-}
-
-target :harness do
-  signature 'sig'
-  check 'projects/harness.rb'
-  check 'projects/harness'
-  ignore 'projects/harness/api', 'projects/harness/tests'
-  library 'json', 'logger', 'optparse', 'time'
-  configure_code_diagnostics do |hash|
-    hash.update(common_diagnostics)
-  end
-end
