+++ workers/ruby/harness/Gemfile
@@ -1,9 +1,11 @@
 source 'https://rubygems.org'
 
 gemspec
-gem 'harness', path: 'harness'
 
 group :development do
+  gem 'grpc-tools', '~> 1.80'
+  gem 'minitest'
+  gem 'rake'
   gem 'rbs', '~> 3.10'
   gem 'rubocop', '~> 1.0'
   gem 'steep', '~> 1.10'
