+++ workers/ruby/harness/Rakefile
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require 'rake/testtask'
+
+Rake::TestTask.new(:test) do |t|
+  t.libs << 'tests'
+  t.pattern = 'tests/test_*.rb'
+end
+
+task default: :test
