+++ workers/ruby/harness/lib/harness.rb
