+++ workers/ruby/harness/lib/harness/client.rb
