+++ workers/ruby/harness/lib/harness/helpers.rb
