+++ workers/ruby/harness/lib/harness/main.rb
