+++ workers/ruby/harness/sig/harness.rbs
