+++ workers/ruby/runner.rb
@@ -1,4 +1,7 @@
+# frozen_string_literal: true
+
+require 'bundler/setup'
+require 'harness'
 require_relative 'kitchen_sink_app'
-require_relative 'projects/harness'
 
 Harness.run(KitchenSinkApp.app) if __FILE__ == $PROGRAM_NAME
